Native java klient for PDP / ABAC
