+++ felles/abac/src/main/java/no/nav/vedtak/sikkerhet/pdp/PdpConsumer.java
@@ -1,8 +1,8 @@
 package no.nav.vedtak.sikkerhet.pdp;
 
 import no.nav.vedtak.sikkerhet.pdp.xacml.XacmlRequestBuilder;
-import no.nav.vedtak.sikkerhet.pdp.xacml.XacmlResponseWrapper;
+import no.nav.vedtak.sikkerhet.pdp.xacml.XacmlResponse;
 
 public interface PdpConsumer {
-    XacmlResponseWrapper evaluate(XacmlRequestBuilder request);
+    XacmlResponse evaluate(XacmlRequestBuilder request);
 }
