+++ felles/abac/src/main/java/no/nav/vedtak/sikkerhet/pdp/xacml/XacmlAttributeSet.java
@@ -1,41 +1,27 @@
 package no.nav.vedtak.sikkerhet.pdp.xacml;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonObjectBuilder;
-
 public class XacmlAttributeSet {
-    private static final String VALUE = "Value";
-    private static final String ATTRIBUTEID = "AttributeId";
-    private static final String ATTRIBUTE = "Attribute";
-
-    private JsonArrayBuilder attributes = Json.createArrayBuilder();
+    private List<XacmlRequest.AttributeAssignment> attributes = new ArrayList<>();
 
     public XacmlAttributeSet addAttribute(String id, String value) {
         Objects.requireNonNull(id, "Name in JsonObject's name/value pair");
         Objects.requireNonNull(value, "Value in JsonObject's name/value pair");
-        attributes.add(createAttribute(id, value));
+        attributes.add(new XacmlRequest.AttributeAssignment(id, value));
         return this;
     }
 
     public XacmlAttributeSet addAttribute(String id, int value) {
         Objects.requireNonNull(id, "Name in JsonObject's name/value pair");
         Objects.requireNonNull(value, "Value in JsonObject's name/value pair");
-        attributes.add(createAttribute(id, value));
+        attributes.add(new XacmlRequest.AttributeAssignment(id, value));
         return this;
     }
 
-    private JsonObjectBuilder createAttribute(String id, String value) {
-        return Json.createObjectBuilder().add(ATTRIBUTEID, id).add(VALUE, value);
-    }
-
-    private JsonObjectBuilder createAttribute(String id, int value) {
-        return Json.createObjectBuilder().add(ATTRIBUTEID, id).add(VALUE, value);
-    }
-
-    JsonObjectBuilder getAttributes() {
-        return Json.createObjectBuilder().add(ATTRIBUTE, attributes);
+    List<XacmlRequest.AttributeAssignment> getAttributes() {
+        return attributes;
     }
 }
