+++ felles/abac/src/main/java/no/nav/vedtak/sikkerhet/pdp/xacml/XacmlRequest.java
@@ -1,4 +1,4 @@
-package no.nav.vedtak.sikkerhet.pdp2.xacml;
+package no.nav.vedtak.sikkerhet.pdp.xacml;
 
 import java.util.List;
 import java.util.Map;
