--- felles/abac/src/main/java/no/nav/vedtak/sikkerhet/pdp2/xacml/Advice.java
@@ -1,8 +1,0 @@
-package no.nav.vedtak.sikkerhet.pdp2.xacml;
-
-public enum Advice {
-    DENY_KODE_6,
-    DENY_KODE_7,
-    DENY_EGEN_ANSATT;
-
-}
