--- felles/abac/src/main/java/no/nav/vedtak/sikkerhet/pdp2/xacml/Category.java
@@ -1,12 +1,0 @@
-package no.nav.vedtak.sikkerhet.pdp2.xacml;
-
-public enum Category {
-    Resource,
-    Action,
-    Environment,
-    AccessSubject,
-    RecipientSubject,
-    IntermediarySubject,
-    Codebase,
-    RequestingMachine;
-}
