+++ felles/sikkerhet/pom.xml
@@ -63,10 +63,6 @@
             <version>0.7.11</version>
         </dependency>
         <dependency>
-            <groupId>jakarta.json</groupId>
-            <artifactId>jakarta.json-api</artifactId>
-        </dependency>
-        <dependency>
             <groupId>no.nav.security</groupId>
             <artifactId>token-client-core</artifactId>
         </dependency>
